[0.4.0] Validate that an associate is a member of the org in some way
With the addition of associate it became possible to make an associate
who is not a member of any squad. Due to the way membership inheritance
is calculated differently for members vs associates (clarified in
README), this resulted in some unintuitive results when this was done at
LR. Fix this by adding a check.